Fix empty grey box for unknown file types (#405)
Previously if files, whose mime-type was not recognized by MimeUtils, were being sent to a Pix-Art-Client only a grey box without text would be shown.
With this commit the general "open file" text is now shown for these files. Previously this text was only shown for files with recognized mime-types, not being specially handled (like images).

Also the Advanced SSA (*.ass) subtitle format was added to MimeUtils as 'text/plain'.